Create 40. Combination Sum II.cpp
https://leetcode.com/problems/combination-sum-ii/description/
As we have to return all the unique combinations, so we’ll sort the array to skip duplicates.
The base case for our helper function is when the target will be 0 then we’ll push our temp to ans.
So here for every recursion, we’ll take a loop to check which could be our next possible element. Here we can have 2 cases:
if(i>index && candidates[i]==candidates[i-1]) we’ve already taken an element with the same value of current element, so simply skip it.
if(candidates[i] > target) as current element is greater than target & the array is in sorted order, so no chance to get any possible solution from the next elements, so break the loop.
We’ll push the current element to our temp vector & call recursion for the next index, also reducing the target value.
After the call, we’ve to pop the current element from temp as we already used this element.
Time complexity: O(2^n * k), where 2^n is the total number of subsequences & k is the average of every combination.